UILD-765: Refactor EditControlPane and add BlockActions component for improved action handling
